# openapi.yaml
openapi: 3.0.3
info:
  title: Groove Garden Backend API
  description: API for authentication, users, playlists, artists, and albums
  version: 1.0.0
servers:
  - url: http://localhost:4000/api
    description: Local development server
  - url: https://groove-garden-backend-api-production.up.railway.app/api
    description: Live server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /auth/login
  schemas:
...
    Playlist:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        created_by_user_id:
          type: integer
    PlaylistSong:
      type: object
      properties:
        playlist_id:
          type: integer
        song_id:
          type: integer
...
    CreatePlaylistRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
        description:
          type: string
    UpdatePlaylistRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
...
    Error:
      type: object
      properties:
        error:
          type: string
...
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ForbiddenError:
      description: Access token is invalid or expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
paths:
...
  /playlists:
    get:
      tags: [Playlists]
      summary: Get all playlists
      responses:
        '200':
          description: List of playlists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Playlist'
    post:
      tags: [Playlists]
      summary: Create playlist
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePlaylistRequest'
      responses:
        '201':
          description: Playlist created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Playlist'
        '400':
          description: Title is required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
  /playlists/my_playlists:
    get:
      tags: [Playlists]
      summary: Get current user's playlists
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: List of user's playlists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Playlist'
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
  /playlists/{id}:
    get:
      tags: [Playlists]
      summary: Get playlist by ID
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Playlist details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Playlist'
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404':
          description: Playlist not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      tags: [Playlists]
      summary: Update playlist
      description: Updates playlist title and/or description; only the owner may update
      operationId: updatePlaylist
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePlaylistRequest'
      responses:
        '200':
          description: Updated playlist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Playlist'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404':
          description: Playlist not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /playlists/{playlist_id}/{song_id}:
    post:
      tags: [Playlists]
      summary: Add song to playlist
      security: [{ bearerAuth: [] }]
      parameters:
        - name: playlist_id
          in: path
          required: true
          schema:
            type: integer
        - name: song_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '201':
          description: Song added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaylistSong'
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Playlist not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags: [Playlists]
      summary: Remove song from playlist
      security: [{ bearerAuth: [] }]
      parameters:
        - name: playlist_id
          in: path
          required: true
          schema:
            type: integer
        - name: song_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Song removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaylistSong'
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...