# src/routes/playlists.js
import express from "express";
import pool from "../db.js";
import authenticateToken from "../middleware/auth.js";

const router = express.Router();

router.get("/", async (req, res) => {
  try {
    const result = await pool.query("SELECT * FROM playlists");
    res.json(result.rows);
  } catch (error) {
    res.status(500).json({ error: "Internal Server Error" });
  }
});

router.get("/my_playlists", authenticateToken, async (req, res) => {
  const userId = req.user.id;
  try {
    const result = await pool.query(
      "SELECT * FROM playlists WHERE created_by_user_id = $1",
      [userId]
    );
    res.json(result.rows);
  } catch (error) {
    res.status(500).json({ error: "Internal Server Error" });
  }
});

router.get("/:id", authenticateToken, async (req, res) => {
  const { id } = req.params;
  try {
    const result = await pool.query(
      `
      SELECT
        *
      FROM
        playlists
      WHERE playlists.id = $1
      `,
      [id]
    );
    if (result.rows.length === 0) {
      return res.status(404).json({ error: "playlist not found" });
    }
    res.json(result.rows[0]);
  } catch (error) {
    res.status(500).json({ error: "Internal Server Error" });
  }
});

router.post("/", authenticateToken, async (req, res) => {
  const userId = req.user.id;
  const { title, description } = req.body;

  if (!title) {
    return res.status(400).json({ error: "Title is required" });
  }

  try {
    const result = await pool.query(
      `INSERT INTO playlists (title, description, created_by_user_id)
       VALUES ($1, $2, $3)
       RETURNING *`,
      [title, description || "None", userId]
    );

    res.status(201).json(result.rows[0]);
  } catch (error) {
    console.error("Error creating playlist:", error);
    res.status(500).json({ error: "Internal Server Error" });
  }
});

router.post("/:playlist_id/:song_id", authenticateToken, async (req, res) => {
  const { playlist_id, song_id } = req.params;
  const userId = req.user.id;
  try {
    const ownerCheck = await pool.query(
      `SELECT created_by_user_id 
       FROM playlists 
       WHERE id = $1`,
      [playlist_id]
    );

    if (ownerCheck.rows.length === 0) {
      return res.status(404).json({ error: "Playlist not found" });
    }

    if (ownerCheck.rows[0].created_by_user_id !== userId) {
      return res
        .status(403)
        .json({ error: "You do not have permission to modify this playlist" });
    }

    const result = await pool.query(
      `INSERT INTO playlist_song (playlist_id, song_id)
       VALUES ($1, $2)
       RETURNING *`,
      [playlist_id, song_id]
    );
    res.status(201).json(result.rows[0]);
  } catch (error) {
    console.error("Error adding song to playlist:", error);
    res.status(500).json({ error: "Internal Server Error" });
  }
});

// PATCH /playlists/:id -> update a playlist (title, description). Only owner can update.
router.patch("/:id", authenticateToken, async (req, res) => {
  const { id } = req.params;
  const userId = req.user.id;
  const { title, description } = req.body;

  // Validate inputs if provided
  if (title !== undefined && (typeof title !== "string" || !title.trim())) {
    return res.status(400).json({ error: "'title' must be a non-empty string if provided" });
  }
  if (description !== undefined && typeof description !== "string") {
    return res.status(400).json({ error: "'description' must be a string if provided" });
  }

  const updates = [];
  const values = [];
  if (title !== undefined) {
    updates.push(`title = $${updates.length + 1}`);
    values.push(title.trim());
  }
  if (description !== undefined) {
    updates.push(`description = $${updates.length + 1}`);
    values.push(description);
  }

  if (updates.length === 0) {
    return res.status(400).json({ error: "At least one field must be provided to update" });
  }

  try {
    // Ensure playlist exists and user is owner
    const ownerCheck = await pool.query(
      `SELECT created_by_user_id FROM playlists WHERE id = $1`,
      [id]
    );
    if (ownerCheck.rows.length === 0) {
      return res.status(404).json({ error: "Playlist not found" });
    }
    if (ownerCheck.rows[0].created_by_user_id !== userId) {
      return res.status(403).json({ error: "You do not have permission to modify this playlist" });
    }

    const query = `UPDATE playlists SET ${updates.join(", ")} WHERE id = $$
${updates.length + 1} RETURNING *`;
    const result = await pool.query(query, [...values, id]);
    return res.status(200).json(result.rows[0]);
  } catch (error) {
    console.error("Error updating playlist:", error);
    return res.status(500).json({ error: "Internal Server Error" });
  }
});

router.delete("/:playlist_id/:song_id", authenticateToken, async (req, res) => {
  const { playlist_id, song_id } = req.params;
  const userId = req.user.id;

  try {
    const ownerCheck = await pool.query(
      `SELECT created_by_user_id 
       FROM playlists 
       WHERE id = $1`,
      [playlist_id]
    );

    if (ownerCheck.rows.length === 0) {
      return res.status(404).json({ error: "Playlist not found" });
    }

    if (ownerCheck.rows[0].created_by_user_id !== userId) {
      return res
        .status(403)
        .json({ error: "You do not have permission to modify this playlist" });
    }

    const result = await pool.query(
      `DELETE FROM playlist_song
       WHERE playlist_id = $1 AND song_id = $2
       RETURNING *`,
      [playlist_id, song_id]
    );

    if (result.rows.length === 0) {
      return res.status(404).json({ error: "Song not found in this playlist" });
    }

    res.status(200).json(result.rows[0]); // Return the deleted row
  } catch (error) {
    console.error("Error removing song from playlist:", error);
    res.status(500).json({ error: "Internal Server Error" });
  }
});

export default router;